moved weapon and projectile to new package
