+++ src/main/java/nl/tudelft/model/LevelFactory.java
@@ -10,6 +10,7 @@ import nl.tudelft.model.bubble.Bubble4;
 import nl.tudelft.model.bubble.Bubble5;
 import nl.tudelft.model.bubble.Bubble6;
 import nl.tudelft.model.pickups.Pickup;
+import nl.tudelft.model.weapon.Projectile;
 import nl.tudelft.semgroup4.Resources;
 
 public class LevelFactory {
