+++ src/main/java/nl/tudelft/model/Player.java
@@ -5,6 +5,7 @@ import java.util.LinkedList;
 import nl.tudelft.model.pickups.Pickup;
 import nl.tudelft.model.pickups.Powerup;
 import nl.tudelft.model.pickups.Powerup.PowerType;
+import nl.tudelft.model.weapon.Weapon;
 import nl.tudelft.semgroup4.Modifiable;
 import nl.tudelft.semgroup4.Resources;
 import nl.tudelft.semgroup4.logger.LogSeverity;
