+++ src/main/java/nl/tudelft/semgroup4/collision/DefaultCollisionHandler.java
@@ -3,14 +3,14 @@ package nl.tudelft.semgroup4.collision;
 import nl.tudelft.model.AbstractGameObject;
 import nl.tudelft.model.Game;
 import nl.tudelft.model.Player;
-import nl.tudelft.model.Projectile;
 import nl.tudelft.model.Wall;
-import nl.tudelft.model.Weapon;
 import nl.tudelft.model.bubble.Bubble;
 import nl.tudelft.model.pickups.Pickup;
 import nl.tudelft.model.pickups.PickupContent;
 import nl.tudelft.model.pickups.Powerup;
 import nl.tudelft.model.pickups.Utility;
+import nl.tudelft.model.weapon.Projectile;
+import nl.tudelft.model.weapon.Weapon;
 import nl.tudelft.semgroup4.logger.LogSeverity;
 import nl.tudelft.semgroup4.util.Audio;
