+++ src/test/java/nl/tudelft/model/GameTest.java
@@ -10,8 +10,10 @@ import static org.mockito.Mockito.when;
 
 import java.util.LinkedList;
 
+import nl.tudelft.model.weapon.Projectile;
 import nl.tudelft.semgroup4.Resources;
 import nl.tudelft.semgroup4.collision.CollisionHandler;
+
 import org.junit.Test;
 import org.newdawn.slick.SlickException;
 import org.newdawn.slick.state.StateBasedGame;
